// frontend/src/api/generated/models/roleEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Dekopen API
 * Authenticated tenant and engine API boundary.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `OWNER` - OWNER
 * * `ESTIMATOR` - ESTIMATOR
 * * `WORKSHOP_MANAGER` - WORKSHOP_MANAGER
 * * `INSTALLER` - INSTALLER
 */
export type RoleEnum = (typeof RoleEnum)[keyof typeof RoleEnum];

export const RoleEnum = {
  OWNER: "OWNER",
  ESTIMATOR: "ESTIMATOR",
  WORKSHOP_MANAGER: "WORKSHOP_MANAGER",
  INSTALLER: "INSTALLER",
} as const;